requirements.txt
git+https://github.com/Rapptz/discord.py
aiomysql
python-dotenv
